refactor: remove unit tests for getters and mutations in the ont pools module
- Deleted the `getters.spec.js` and `mutations.spec.js` files as part of a cleanup process.
- Updated the `ONTSampleIndex.spec.js` to use a new mounting function with store integration.
